+++ cloud-components/label_nodes.sh
@@ -6,7 +6,7 @@
 # Author: Feng Li
 # email: fengggli@yahoo.com
 
-worker_nodes=($(kubectl get nodes -l  magnum.openstack.org/role=worker -o jsonpath={.items[*].metadata.name}))
+worker_nodes=($(kubectl get nodes --selector='!node-role.kubernetes.io/master' -o jsonpath={.items[*].metadata.name}))
 
 i=0
 for node in ${worker_nodes[*]}
@@ -14,5 +14,6 @@ do
 echo setting for node ${i}}
 #kubectl label nodes ${worker_nodes[i]}  minion-idx=${i}
 kubectl label --overwrite nodes $node  minion-idx=${i}
+kubectl label --overwrite nodes $node  is-minion=true
 i=$((i+1))
 done