+++ tests/test-iperf3/collect_result.sh
@@ -1,0 +1,16 @@
+#!/usr/bin/env bash
+
+# First created: 
+# Last modified: 2018 Jan 26
+
+# Author: Feng Li
+# email: fengggli@yahoo.com
+JOBID=$1
+
+for logfile in results/$JOBID/log.*; do
+  echo `basename $logfile`:
+  cat $logfile| grep sender|tail -1 ;
+done
+
+
+